New build

+++ app/Http/Controllers/AdditionalCostController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\AdditionalCostType;
 use App\Helper\Helper;
 use App\Http\Requests\AdditionalCostRequest;
 use App\Http\Resources\AdditionalCostCollection;
 use App\Models\AdditionalCost;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
@@ -13,6 +15,8 @@ class AdditionalCostController extends Controller
 {
     public function index()
     {
+        $this->authorize('showAdditionalCost',User::class);
+
         $requestParam = \request()->all('search', 'trashed');
         return Inertia::render('AdditionalCost/Index', [
             'filters' => $requestParam,
@@ -27,6 +31,8 @@ class AdditionalCostController extends Controller
 
     public function create()
     {
+        $this->authorize('createAdditionalCost',User::class);
+
         return Inertia::render('AdditionalCost/Create',[
             ...Helper::messArray()
         ]);
@@ -34,10 +40,14 @@ class AdditionalCostController extends Controller
 
     public function store(AdditionalCostRequest $request)
     {
-        AdditionalCost::create(
+        $this->authorize('createAdditionalCost',User::class);
+
+        $additional = AdditionalCost::create(
             $request->validated()
         );
-
+        if($request->status === AdditionalCostType::APPROVED){
+            $additional->mess()->decrement('deposit',$additional->amount);
+        }
         return to_route('additional.index');
     }
 
@@ -49,23 +59,35 @@ class AdditionalCostController extends Controller
 
     public function edit(AdditionalCost $additional)
     {
+        $this->authorize('editAdditionalCost',User::class);
+
         return Inertia::render('AdditionalCost/Edit', [
-            'additional' => $additional,
-            ...Helper::messArray()
+            'additional' => $additional
         ]);
     }
 
     public function update(AdditionalCostRequest $request, AdditionalCost $additional)
     {
+        $this->authorize('editAdditionalCost',User::class);
+
+        $additional->mess()->increment('deposit',$additional->amount);
+
         $additional->update(
             $request->validated()
         );
+
+        if($request->status === AdditionalCostType::APPROVED){
+            $additional->mess()->decrement('deposit',$additional->amount);
+        }
 
         return to_route('additional.index');
     }
 
     public function destroy(AdditionalCost $additional)
     {
+        $this->authorize('deleteAdditionalCost',User::class);
+
+        $additional->mess()->increment('deposit',$additional->amount);
         $additional->delete();
 
         return to_route('additional.index');
