+++ app/Http/Controllers/DepositController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Enums\DepositStatus;
 use App\Helper\Helper;
 use App\Http\Requests\DepositRequest;
 use App\Http\Requests\WithDrawRequest;
@@ -16,6 +17,8 @@ class DepositController extends Controller
 {
     public function index()
     {
+        $this->authorize('showDeposit',Deposit::class);
+
         $requestParam = \request()->all('search', 'trashed');
         return Inertia::render('Deposit/Index', [
             'filters' => $requestParam,
@@ -27,7 +30,8 @@ class DepositController extends Controller
                         $query->select(
                             'user_id',
                             DB::raw('SUM(CASE WHEN status = 1 THEN amount ELSE 0 END) as deposit_amount'),
-                            DB::raw('SUM(CASE WHEN status = 3 THEN amount ELSE 0 END) as withdraw_amount')
+                            DB::raw('SUM(CASE WHEN status = 3 THEN amount ELSE 0 END) as withdraw_amount'),
+                            DB::raw('SUM(CASE WHEN status = 0 THEN amount ELSE 0 END) as pending_amount'),
                         )->groupBy('user_id');
                     })
                     ->select('id', 'first_name', 'last_name')
@@ -40,6 +44,8 @@ class DepositController extends Controller
 
     public function create()
     {
+        $this->authorize('createDeposit',Deposit::class);
+
         return Inertia::render('Deposit/Create', [
             ...Helper::usersArray()
         ]);
@@ -47,20 +53,28 @@ class DepositController extends Controller
 
     public function store(DepositRequest $request)
     {
-        Deposit::create(
+        $this->authorize('createDeposit',Deposit::class);
+
+
+        $deposit = Deposit::create(
             $request->validated()
         );
+
+        $deposit->user()->increment('deposit', $deposit->amount);
+        $deposit->mess()->increment('deposit', $deposit->amount);
+
 
         return redirect()->back()->with('success', 'New Deposit Added');
     }
 
     public function show($userId)
     {
-        $queryModel = Deposit::whereUserId($userId)->where('status', '!=', 0);
+        $this->authorize('showDeposit',Deposit::class);
+
+
         return Inertia::render('Deposit/Show', [
             'user' => User::find($userId),
             'approvedDeposit' => Deposit::whereUserId($userId)->where('status', '!=', 0)->get(),
-            'total' => Deposit::whereUserId($userId)->whereIn('status', [1])->sum('amount'),
             'pendingDeposit' => Deposit::whereUserId($userId)->whereStatus(0)->get(),
         ]);
 
@@ -69,6 +83,9 @@ class DepositController extends Controller
 
     public function edit(Deposit $deposit)
     {
+        $this->authorize('editDeposit',Deposit::class);
+
+
         return Inertia::render('Deposit/Edit', [
             'deposit' => $deposit,
             ...Helper::usersArray()
@@ -77,6 +94,8 @@ class DepositController extends Controller
 
     public function update(DepositRequest $request, Deposit $deposit)
     {
+        $this->authorize('editDeposit',User::class);
+
         $deposit->update(
             $request->validated()
         );
@@ -86,6 +105,10 @@ class DepositController extends Controller
 
     public function destroy(Deposit $deposit)
     {
+        $this->authorize('deleteDeposit',Deposit::class);
+
+        $deposit->user()->decrement('deposit', $deposit->amount);
+        $deposit->mess()->decrement('deposit', $deposit->amount);
         $deposit->delete();
 
         return redirect()->back()->with('success', 'Deposit deleted successfully');
@@ -100,22 +123,40 @@ class DepositController extends Controller
 
     public function accept(Deposit $deposit)
     {
-        $deposit->status = 1;
+        $this->authorize('approveDeposit',Deposit::class);
+
+
+        $deposit->status = DepositStatus::APPROVED;
         $deposit->save();
+        $deposit->user()->increment('deposit', $deposit->amount);
+        $deposit->mess()->increment('deposit', $deposit->amount);
         return redirect()->back()->with('success', 'Deposit approved');
     }
 
     public function reject(Deposit $deposit)
     {
+        $this->authorize('rejectDeposit',Deposit::class);
+
+
         $deposit->forceDelete();
         return redirect()->back()->with('success', 'Deposit Deleted');
     }
 
     public function withdraw(WithDrawRequest $request)
     {
-        Deposit::create(
+        $this->authorize('withdrawDeposit',Deposit::class);
+
+        $user = User::find($request->user_id);
+
+        if($request->amount > $user->deposit){
+            return back()->with('errors','Withdraw amount is greater than deposit amount');
+        }
+        $deposit = Deposit::create(
             $request->validated()
         );
+
+        $user->decrement('deposit', $deposit->amount);
+        $deposit->mess()->decrement('deposit', $deposit->amount);
 
         return redirect()->back()->with('success', 'New Withdrawal added');
     }
